user list role editing  auth

+++ app/Http/Controllers/UserController.php
@@ -8,16 +8,19 @@ use App\Models\User;
 class UserController extends Controller
 {
     public function index(){
+        if(!auth()->user() || auth()->user()->role !== 's') abort(401);
         return view('users', ['users' => User::get()]);
     }
 
     public function update(Request $request, User $user){
+        if(!auth()->user() || auth()->user()->role !== 's') return ['success' => false, 'msg' => 'Unauthorized!'];
         $user->role = $request->input('role', $user->role);
         $user->save();
         return ['success' => true, 'users' => User::get()];
     }
 
     public function delete(Request $request, User $user){
+        if(!auth()->user() || auth()->user()->role !== 's') return ['success' => false, 'msg' => 'Unauthorized!'];
         $user->delete();
         return ['success' => true];
     }
